Fix mobile build due to location changes
