+++ packages/common/src/services/index.ts
@@ -13,3 +13,4 @@ export * from './solana-client'
 export * from './cognito'
 export * from './track-download'
 export * from './oauth'
+export * from './location'
