+++ packages/web/src/common/store/pages/signon/sagas.js
@@ -13,7 +13,8 @@ import {
   getContext,
   settingsPageActions,
   MAX_HANDLE_LENGTH,
-  PushNotificationSetting
+  PushNotificationSetting,
+  getCityAndRegion
 } from '@audius/common'
 import { push as pushRoute } from 'connected-react-router'
 import { isEmpty } from 'lodash'
@@ -38,7 +39,6 @@ import { fetchUserByHandle, fetchUsers } from 'common/store/cache/users/sagas'
 import { processAndCacheUsers } from 'common/store/cache/users/utils'
 import * as confirmerActions from 'common/store/confirmer/actions'
 import { confirmTransaction } from 'common/store/confirmer/sagas'
-import { getCityAndRegion } from 'services/Location'
 import { UiErrorCode } from 'store/errors/actions'
 import { setHasRequestedBrowserPermission } from 'utils/browserNotifications'
 import { isValidEmailString } from 'utils/email'
